tests/functional/requirements.txt
pytest==7.4.4
pytest-asyncio==0.12.0
aiohttp==3.7.2
pydantic==2.1.1
pydantic-settings==2.0.1
python-dotenv==1.0.0
elasticsearch[async]==7.9.1
redis==4.1.0
